Use ActiveSupport::Configurable to set poll time and cache expiry.
Use global config instead of making individual classes configurable.
More maintainable if there are additional configuration options as
everything will be in one place.

Parse the poll duration and convert to seconds when running the client.
Remove extra arguments needed previously for passing poll time to the
Client.
Update tests to check if sleep time is the same as the configured time.

Add tests to ensure that configuration is set when CLI runs.
Use the configured value for image lifetime in DockerContainer.